Added only json file
Added only json file, without any changes to the original code.

The json file is still to be integrated in the app, if you want to integrate it you can find my code in the previous commit or let me know if you want a more optimized version to implement it on the original code.